chore(ui): move styled input
